==> 4_NFL_Predictor_Prediction_Generator.ipynb <==
from nfl import process #saved as a package the make_feature function defined in previous notebook
import pandas as pd
import numpy as np
from sklearn.preprocessing import Imputer, RobustScaler, StandardScaler

from sklearn.linear_model import LinearRegression, ElasticNetCV

from sklearn.grid_search import GridSearchCV

from time import ctime

from sklearn.cross_validation import cross_val_score

import seaborn as sns

import matplotlib.pyplot as plt

from sklearn.cross_validation import train_test_split, ShuffleSplit
from sklearn.ensemble import ExtraTreesRegressor

import statsmodels.formula.api as smf
from statsmodels.api import add_constant

from sklearn.metrics import mean_squared_error, r2_score 

from sklearn.svm import SVR, LinearSVR

import scipy.stats as stats
import math

import time

from sklearn.linear_model import Lasso, Ridge, RidgeCV


%matplotlib inline

### The general game plan is as follows. Start with several years of data for each position, and generate week by week data for the past by re-training the model on all of the prior data.

df = pd.read_csv("~/ProjectData/NFL/all_data_10_13_v2.csv", low_memory=False)

wr = process.make_feature("WR", 2010, df = df)
wr.head(3)

qb = process.make_feature("QB", 2010, df = df)
qb.head(3)

te = process.make_feature("TE", 2010, df = df)
te.head(3)

rb = process.make_feature("RB", 2010, df = df)
rb.head(3)

wr_X, wr_y, wr_frame = process.crop_and_separate(0, 1, wr, "data")

qb_X, qb_y , qb_frame = process.crop_and_separate(0, 1, qb, "data")

rb_X, rb_y, rb_frame = process.crop_and_separate(0, 1, rb, "data")

te_X, te_y, te_frame = process.crop_and_separate(0, 1, te, "data")

###### PROTOTYPE THE PREDICTION LOOP

#the loop will use the frame to figure out the X weights and the indexes on which to fit/predict the X and y
X, y, frame = process.crop_and_separate(0, 2, wr, "data")
frame.head()

#define model, using the RF with the attributes that I grid searched for in notebook 3
et4 = ExtraTreesRegressor(max_depth=7,
                          min_samples_leaf= 30, 
                          min_weight_fraction_leaf=0.13740419750125152, 
                          n_estimators=50, n_jobs = -1)



for year in set(frame.seas): print year

sy = 2014 #years to predict forward from
cy = 2015 #looping current year of prediction
cw = 2 # looping current week of prediction
player = "TS-2300"

for week in set(frame[frame.seas == sy].wk): print week

#define prediction index
predict_rows = frame[(frame.seas == cy) & (frame.wk == cw) & (frame.player == player)].index

#define training index
train_rows = frame[(((frame.seas == cy) & (frame.wk <=cw)) | (frame.seas < cy)) ].index

#### Another thing to do here is to, compute a sample weighting vector, but i want to time it as well to see if its realistic to re-fit per player per week. The weighting would apply a multiplication factor to the players score and weight by ratio of scores so it would be two fold, favor the higher perfomances since that is what I am targeting and also add an extra weight to each particular 

frame["weight"] = frame.target

%%timeit

#a = time.time()


frame.ix[frame.player == player,"weight"] = frame.ix[frame.player == player,"weight"] *10 
frame.weight = frame.weight/sum(frame.weight)
#print sum(frame.weight)

et4.fit(X.ix[train_rows], y.ix[train_rows], sample_weight= np.array(frame.weight.ix[train_rows]))
x = et4.predict(X.ix[predict_rows])
#b = time.time()

#print b-a

##### PREDICTION FUNCTION

def generate_prediction(X, y, frame, sy, model, w=10):
    """generates predictions by player by week with the option to increase the weight of the players
    historic performances. w = 1 would weight every players game by the amount of fp scored without
    increasing the weight of each player's own games. w=0 does not apply any sample weight.
    
    frame = dataframe resulting from the make feature function
    sy = starting year to predict from
    w = factor to apply to each players scores when weighting samples
    model = predictor with a fit/predict method"""
    yr_set = pd.Series([_ for _ in set(frame.seas)])
    cycle_yrs = yr_set[yr_set>=sy]
    for year in cycle_yrs:
        week_set = pd.Series([_ for _ in set(frame.wk[frame.seas==year])])
        for week in week_set:
            player_set = pd.Series([_ for _ in set(frame.player[(frame.seas == year) & (frame.wk == week)])])
            for player in player_set:
                predict_rows = frame[(frame.seas == year) & (frame.wk == week) & (frame.player == player)].index
                train_rows = frame[(((frame.seas == year) & (frame.wk <=week)) | (frame.seas < year)) ].index
                
                if w == 0:
                    model.fit(X.ix[train_rows], y.ix[train_rows])
                else:
                    print player, year, week
                    frame["weight"] = frame.target
                    frame.ix[frame.player == player,"weight"] = frame.ix[frame.player == player,"weight"] *w
                    frame.weight = frame.weight/sum(frame.weight)
                    model.fit(X.ix[train_rows], y.ix[train_rows], sample_weight= np.array(frame.weight.ix[train_rows]))
                
                frame.ix[predict_rows,"prediction"] = model.predict(X.ix[predict_rows])
                
            print "Week ", week, " year ", year, " is done."
        
    return frame
    
    


lr = Lasso()

a= time.time()
qb_result = generate_prediction(qb_X, qb_y, qb_frame, 2014, lr, w=0)
print time.time() - a

a= time.time()
wr_result = generate_prediction(wr_X, wr_y, wr_frame, 2014, lr, w=0)
print time.time() - a

a= time.time()
rb_result = generate_prediction(rb_X, rb_y, rb_frame, 2014, lr, w=0)
print time.time() - a

a= time.time()
te_result = generate_prediction(te_X, te_y, te_frame, 2014, lr, w=0)
print time.time() - a

#save the prediction frames down so i can run some analysis in the next notebook

qb_result["pos"] = "QB"
wr_result["pos"] = "WR"
rb_result["pos"] = "RB"
te_result["pos"] = "TE"

grab_cols = ["target", "prediction", "gid", "player",  "pos", "fname", "lname", "seas", "wk"]

qb_result.head(1)



final = qb_result[grab_cols].append(wr_result[grab_cols]).append(rb_result[grab_cols]).append(te_result[grab_cols])

final.to_csv("predictions_10_20.csv")

test = final[["target", "prediction"]].dropna()

r2_score(test.target, test.prediction)

#### The results of the RF are not here but I was not getting a good R2 score (close to 0) which may have been due to overfitting in the third notebook but it looks like a Lasso model works failry well here across all positions so I will move on to the next step which is the line up optimization.
