==> 5_NFL_Predictor_Weekly_Optimization_Final.ipynb <==
from nfl import process #saved as a package the make_feature function defined in previous notebook
import pandas as pd
import numpy as np
from sklearn.preprocessing import Imputer, RobustScaler, StandardScaler

from sklearn.linear_model import LinearRegression, ElasticNetCV

from sklearn.grid_search import GridSearchCV

from time import ctime

from sklearn.cross_validation import cross_val_score

import seaborn as sns

import matplotlib.pyplot as plt

from sklearn.cross_validation import train_test_split, ShuffleSplit
from sklearn.ensemble import ExtraTreesRegressor

import statsmodels.formula.api as smf
from statsmodels.api import add_constant

from sklearn.metrics import mean_squared_error, r2_score 

from sklearn.svm import SVR, LinearSVR

import scipy.stats as stats
import math

import time

import collections
import functools 

%matplotlib inline


#a memoized class taken from https://wiki.python.org/moin/PythonDecoratorLibrary#Memoize
#this caches results of a function and does not re-compute if not needed
class memoized(object):
   '''Decorator. Caches a function's return value each time it is called.
   If called later with the same arguments, the cached value is returned
   (not reevaluated).
   '''
   def __init__(self, func):
      self.func = func
      self.cache = {}
   def __call__(self, *args):
      if not isinstance(args, collections.Hashable):
         # uncacheable. a list, for instance.
         # better to not cache than blow up.
         return self.func(*args)
      if args in self.cache:
         return self.cache[args]
      else:
         value = self.func(*args)
         self.cache[args] = value
         return value
   def __repr__(self):
      '''Return the function's docstring.'''
      return self.func.__doc__
   def __get__(self, obj, objtype):
      '''Support instance methods.'''
      return functools.partial(self.__call__, obj)

preds = pd.read_csv("predictions_10_20.csv") #import predictions generated in notebook #3

preds.dropna(inplace = 1)

preds = preds[preds.wk < 18] #drop non reg season weeks

preds.head()

salaries = pd.read_csv("/Users/egorkharakozov/ProjectData/NFL/player_salaries.csv", header = None)

#file from rotoguru http://rotoguru1.com/cgi-bin/fyday.pl?week=1&year=2014&game=dk&scsv=1
#downloaded with importIO free scraper and converted using text to columns in excel.
# The data didnt really require a custom scraper built through python.



salaries.head()

#reshape and process the salary data to fit each player on a row. The data came in with each week being a row.

processed_sals = pd.DataFrame()

for i in range(0,34):
    temp = salaries.ix[i,:].reshape(438,9)
    temp = pd.DataFrame(temp)
    temp["week"] = i//2+1
    
    processed_sals = processed_sals.append(temp)


    
processed_sals.columns = ["seas", "game_id", "pname", "pos", "home_team", "h/a", "opp_team", "pts", "sal", "wk"]

processed_sals.head()

processed_sals.sal = [str(_).split()[0] for _ in processed_sals.sal]

processed_sals.sal = processed_sals.sal.astype("float64")

processed_sals.dropna(inplace = 1)

preds["pname"] = preds.lname +", " + preds.fname

preds.head()

test2 = processed_sals.merge(preds, how = "left", on = ["pname", "seas", "wk", "pos"]) 
#merge the predictions and salary data

test2.head()

### In the below section I mock up the optimization script. The first function will reduce the weekly field of players, and the second function will optimize based on salary and roster limit constraints using a modified dynamic programming function similiar to the solution to the famous knapsack problem. 



poslist = {"QB":1, "WR":4, "RB":3, "TE":2}

def crop_weak_players(df, poslist, filtering_value = "prediction"):
    
    """returns an index on non-reduntant players by testing for 
    a dominant condition where player A is cheaper than player B and
    player A is projected for more points than player B and therefore player B is eliminated from the list.
    To avoid eliminating too many players, this function also checks the position list to see how many
    players must exist before eliminating
    parameters:
    df - a dataframe containing pos, sal, prediction, and name columns
    poslist - a dictionary for example  {"QB":1, "WR":4, "RB":3, "TE":2} showing max position limits in lineup
    """
    
    reduced_index = []
    for k, v in poslist.iteritems():
        for i1, v1 in df[df.pos==k].iterrows():
            count = 0
            for i2, v2 in df[df.pos==k].iterrows():
                if v2.name == v1.name:
                    continue
                elif (v2.sal <= v1.sal) & (v2[filtering_value] >= v1[filtering_value]):
                    count += 1
            if count >= v:
                pass
            else:
                reduced_index.append(i1)

    return reduced_index


#try the knapsack approach below here


grabcols = ["seas", "wk","pos", "pname", "sal", "target", "prediction", "player"]

test2 = test2[grabcols]


#here i clean up the data to trim down to a dataset per week that i will then run the knapsack optimization on

test2 = test2[(test2.wk == 16) & (test2.seas == 2015)]
tot_sal = 500 -  np.round(np.mean(test2.sal[test2.pos == "Def"]),-2)/100
test2 = test2[test2.pos != "Def"]
print test2.pos.value_counts()

test2.sal[test2.sal < 2000] = 0 #2k is min salary so anything under that is a typo or bad data point
test2.sal.replace(0, int(np.round(np.mean(test2.sal),-2)), inplace = 1)

test2.sal = [int(_)/100 for _ in test2.sal]
test2.dropna(inplace = 1)
test2.prediction = [int(np.rint(_*100)) for _ in test2.prediction]
#test2 = test2.sort_values(by = ["prediction"], ascending=False)

print tot_sal
test2.describe()

reduced_index = crop_weak_players(test2, poslist)
test2 = test2.ix[reduced_index]


test2.reset_index(drop = 1, inplace= 1)

#test2

#add weights for positions as that will be used as a weight
test2["pos2"] = test2["pos"]
test2 = pd.get_dummies(test2, columns = ["pos2"])


#add a weight for flex (rb + wr)
test2["tot"] = np.sum(test2.ix[:,-4:], axis = 1)
test2["rbwr"] = np.sum(test2.ix[:,["pos2_RB", "pos2_WR"]], axis = 1)
test2["rbte"] = np.sum(test2.ix[:,["pos2_RB", "pos2_TE"]], axis = 1)
test2["wrte"] = np.sum(test2.ix[:,["pos2_WR", "pos2_TE"]], axis = 1)
test2.head()

#convert to integers as that will create discrete values and take advantage of the memoized construct

for col in ["pos2_QB", "pos2_RB", "pos2_WR", "pos2_TE", "tot", "rbwr", "rbte", "wrte"]:
    test2.ix[:,col] = test2.ix[:,col].astype("int")


def knapsack2(items, maxweight, poslist, rbwr = 6, rbte = 4, wrte = 5, roster = 8, value = "prediction"):
    """
    Solve the knapsack problem by finding the most valuable
    subsequence of `items` subject that weighs no more than
    the weight constraints.

    `items` is a df returned by the filter_weak_players with added columns for constraints.

    `maxweight` is the salary limit
    
    `poslist` is a dictionary with positional max size limits
    
    `flex` is the max total rb + wr allowed
    
    `roster` is total roster size

    This returns a dataframe with the optimal lineup selected and the total predicted score
    """
    
    grabcols = ["pname", "pos", "sal", "pos2_QB", "pos2_WR", "pos2_RB","pos2_TE", "rbwr", "rbte", "wrte", "tot", value]
    items = items[grabcols]
    # Return the value of the most valuable subsequence of the first i
    # elements in items whose weights sum to no more than j.
    @memoized
    def bestvalue(i, j, k, l, m, n, o, p, q, r):
        if i == 0: return 0
        _, pos, weight, qbw, wrw, rbw, tew, fx1, fx2, fx3, tw, val = items.ix[i - 1, :]
        if ((weight > j) | (qbw > k) | (wrw > l) | (rbw > m) | (tew > n )\
            | (fx1 > o) |(fx2 > p) |(fx3 > q) | (tw > p)) :
            return bestvalue(i - 1, j, k, l, m, n, o, p, q, r)
        else:
            return max(bestvalue(i - 1, j, k, l, m, n, o, p, q, r),
                       bestvalue(i - 1, j - weight, k-qbw, l-wrw, m-rbw, n-tew,\
                                 o-fx1, p - fx2, q- fx3, r - tw) + val)

    j = maxweight
    k = poslist["QB"]
    l = poslist["WR"]
    m = poslist["RB"]
    n = poslist["TE"]
    o = rbwr
    p = rbte
    q = wrte
    r = roster 
    ### total roster of 9 constraint, would have to be manually changed if optimizing for other websites
    
    
    result = pd.DataFrame(columns = ["pname", "pos", "sal", "pos2_QB", "pos2_WR",
                                     "pos2_RB","pos2_TE", "rbwr", "rbte", "wrte", "tot", value])
    for i in xrange(len(items), 0, -1):
        if bestvalue(i, j, k, l, m, n, o, p, q, r) != bestvalue(i - 1, j, k, l, m, n, o, p, q, r):
            result =result.append(items.ix[i - 1, :])
            j -= items.ix[i - 1, :].sal
            k -= items.ix[i - 1, :].pos2_QB
            l -= items.ix[i - 1, :].pos2_WR
            m -= items.ix[i - 1, :].pos2_RB
            n -= items.ix[i - 1, :].pos2_TE
            o -= items.ix[i - 1, :].rbwr
            p -= items.ix[i - 1, :].rbte
            q -= items.ix[i - 1, :].wrte
            r -= items.ix[i - 1, :].tot
           
    return bestvalue(len(items), maxweight, poslist["QB"], 
                     poslist["WR"], poslist["RB"], poslist["TE"], rbwr, rbte, wrte, roster)/100.0, result

a = time.time()
score2 , lineup2 = knapsack2(test2, tot_sal, poslist)

print time.time() - a

lineup2

score2

sum(lineup2.prediction/100.0)

### as a quick sanity test, setting a trivially large salary constraint 
#should return just the maximum predicted lineup.
#which it does

score2 , lineup3 = knapsack2(test2, 99999999, poslist)
lineup3

sum(lineup2.sal) +500 - tot_sal

### Now that I have a function that trims each weekly player set and generates the optimal lineup, I can create a loop to generate the optimized line ups per week and see how they perform against scores likely that would have needed to cash. The best source I have found so far on average scores needed is here 
https://rotogrinders.com/articles/what-s-needed-to-win-on-draftkings-368296
### and this article stipulates that although the average score is 129, because the distribution has fatter tails, (platykurtic) in stats speak, the last winning score is closer to the 119 mark. Given that this article is from 2014 we can split the difference and test the 34 predicted weeks against both levels and the midpoint to gauge the model's performance.


#first I create a df with all of the predicted points and salaries in one frame:

all_data = processed_sals.merge(preds, how = "left", on = ["pname", "seas", "wk", "pos"])

all_data.head()

grabcols = ["seas", "wk","pos", "pname", "sal", "target", "prediction", "player","pts"]

poslist = {"QB":1, "WR":4, "RB":3, "TE":2}


all_data = all_data[grabcols]

### PART ONE GENERATE THE OPTIMAL LINE UP FOR EACH WEEK USING THE PREDICTION MODEL RESULTS

optimized_lineups = pd.DataFrame()

for year in [2014, 2015]:
    for week in set(all_data.wk):
        mask = ((all_data.seas == year) & (all_data.wk == week))
        temp = all_data[mask]
        def_sal = np.round(np.mean(temp.sal[temp.pos == "Def"]),-2)/100
        def_pts = np.round(np.mean(temp.pts[temp.pos == "Def"]),-2)/100
        tot_sal = 500 - def_sal
        temp = temp[temp.pos != "Def"]
        temp.sal[temp.sal <2000] = 0
        temp.sal.replace(0, int(np.round(np.mean(temp.sal),-2)), inplace = 1)
        temp.sal = [int(_)/100 for _ in temp.sal]
        temp.dropna(inplace = 1)
        temp.prediction = [int(np.rint(_*100)) for _ in temp.prediction]
        temp.target = [int(np.rint(_*100)) for _ in temp.target]
        
        reduced_index = crop_weak_players(temp, poslist)
        
        
        temp = temp.ix[reduced_index]
        temp.reset_index(drop =1 , inplace = 1)
        temp["pos2"] = temp["pos"]
        temp = pd.get_dummies(temp, columns = ["pos2"])
        temp["tot"] = np.sum(temp.ix[:,-4:], axis = 1)
        
        
        temp["rbwr"] = np.sum(temp.ix[:,["pos2_RB", "pos2_WR"]], axis = 1)
        temp["rbte"] = np.sum(temp.ix[:,["pos2_RB", "pos2_TE"]], axis = 1)
        temp["wrte"] = np.sum(temp.ix[:,["pos2_WR", "pos2_TE"]], axis = 1)
                #convert to integers as that will create discrete values and take advantage of the memoized construct
        for col in ["pos2_QB", "pos2_RB", "pos2_WR", "pos2_TE", "tot", "rbwr", "rbte", "wrte"]:
            temp.ix[:,col] = temp.ix[:,col].astype("int")
            
        _ , lineup = knapsack2(temp, tot_sal, poslist)
        
        
        lineup["seas"] = year
        lineup["wk"] = week
        lineup["def_sal"] = def_sal/8 
        lineup["def_pts"] = def_pts/8.0
        #add one eights of the def_sal to each week so that when i sum i can double check each week is under limit
        
        optimized_lineups = optimized_lineups.append(lineup)
        print "Done with year {} week {}".format(year, week)
        
        
        

#merge with preds to get the actual scores in
optimized_lineups = optimized_lineups.merge(preds[["pname", "seas", "wk", "target"]], on= ["pname", "seas", "wk"])

optimized_lineups["prediction"] = optimized_lineups["prediction"] / 100.0

optimized_lineups["tot_sal"] = optimized_lineups.sal + optimized_lineups.def_sal
optimized_lineups["tot_score"] = optimized_lineups.def_pts + optimized_lineups.target

#optimized_lineups.drop(258,inplace = 1) 
#there are two RBs named david johnshon on this week so when i merge to bring in the target score i get a dupe


optimized_lineups[(optimized_lineups.wk == 16) & (optimized_lineups.seas == 2015)]

a = optimized_lineups.pivot_table(values = ["prediction", "tot_score", "tot_sal"], index = ["seas", "wk"], aggfunc = "sum")

len(a[a.tot_score>124]) #number of weeks that would have won given the 119, 124, 129 point thresholds
# I derive them from this article and take the midpoint between mean 129 and projected 119. 
#https://rotogrinders.com/articles/what-s-needed-to-win-on-draftkings-368296
    

len(a[a.tot_score > 119])

len(a[a.tot_score > 129])

-(50*34) + 90*21  #amount of money won over 2 years playing a $50 50/50 once a week 
#(of course assuming no reinvestment and flat hurdle of 119 to win and also an average defense)
